# api/swagger/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/development.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "DevelopmentService"
    },
    {
      "name": "GatewayService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/gateway/dev/data-decrypt": {
      "post": {
        "operationId": "DevelopmentService_DecryptData",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gatewayDevDecryptDataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/gatewayDevDecryptDataRequest"
            }
          }
        ],
        "tags": [
          "DevelopmentService"
        ]
      }
    },
    "/api/gateway/dev/data-encrypt": {
      "post": {
        "operationId": "DevelopmentService_EncryptData",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gatewayDevEncryptDataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/gatewayDevEncryptDataRequest"
            }
          }
        ],
        "tags": [
          "DevelopmentService"
        ]
      }
    },
    "/api/gateway/dev/decrypt-key-id": {
      "post": {
        "operationId": "DevelopmentService_DecryptSecretSessionId",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gatewayDecryptSecretSessionIdResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/gatewayDecryptSecretSessionIdRequest"
            }
          }
        ],
        "tags": [
          "DevelopmentService"
        ]
      }
    },
    "/api/gateway/dev/key": {
      "get": {
        "operationId": "DevelopmentService_GenerateSecretKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gatewayGenerateSecretKeyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "DevelopmentService"
        ]
      }
    },
    "/api/gateway/dev/secret-session": {
      "get": {
        "operationId": "DevelopmentService_SetSecretSessionTimeout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "timeout",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DevelopmentService"
        ]
      }
    },
    "/api/gateway/init": {
      "post": {
        "operationId": "GatewayService_InitSecretKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gatewayEncryptedGatewayResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/gatewayEncryptedGatewayRequest"
            }
          }
        ],
        "tags": [
          "GatewayService"
        ]
      }
    }
  },
  "definitions": {
    "gatewayDecryptSecretSessionIdRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        },
        "secretKey": {
          "type": "string"
        }
      }
    },
    "gatewayDecryptSecretSessionIdResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        }
      }
    },
    "gatewayDevDecryptDataRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        },
        "secretKey": {
          "type": "string"
        }
      }
    },
    "gatewayDevDecryptDataResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        }
      }
    },
    "gatewayDevEncryptDataRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        },
        "secretKey": {
          "type": "string"
        }
      }
    },
    "gatewayDevEncryptDataResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        }
      }
    },
    "gatewayEncryptedGatewayRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        }
      }
    },
    "gatewayEncryptedGatewayResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        }
      }
    },
    "gatewayGenerateSecretKeyResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        },
        "secretKey": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}